# docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: EDY HelpDesk Internal API
  version: 1.0.0
  description: Local-first, cookie-session API. Examples are synthetic and contain no credentials.
servers:
  - url: /api/v1
paths:
...
  /auth/login:
    post: { summary: Create a rotated server-side session, responses: { '200': { description: Authenticated }, '401': { description: Invalid credentials }, '429': { description: Rate limited } } }
  /auth/logout:
    post: { summary: Revoke the current session, security: [{ cookieSession: [] }], responses: { '204': { description: Revoked } } }
  /auth/me:
    get: { summary: Read current session, security: [{ cookieSession: [] }], responses: { '200': { description: Session }, '401': { description: Unauthenticated } } }
...
components:
  securitySchemes:
    cookieSession: { type: apiKey, in: cookie, name: edy_session }
    csrf: { type: apiKey, in: header, name: x-csrf-token }